pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["skymagenta <jokero.sky1@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pyyaml = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
